+++ src/js/popup.js
@@ -35,6 +35,7 @@ $(function () {
       $('.list-group').html('')
       return
     }
+    $('.list-group').html('<div id="loading-spinner"><div></div></div>')
     chrome.runtime.sendMessage({action: 'msg.jsonrequest', url: tmpl + searchStr}, function (data) {
       var results = data.message.result.results
       var links = []