+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.0.13] - 2018-10-15
 ### Added
  - Add support for document links with multiple categories / languages / recipient countries ([#69](https://github.com/pwyf/iati-org-viz/issues/69))
  - Add loading spinner to popup